#352 Update to invoiveshow page
